+++ CHANGELOG.md
@@ -1,9 +1,9 @@
 # Changelog
 
-## Latest Changes
 
 ### 💡 Hack Examples
-* 2025-01-21 - eg: use sed to do substitution
+* 2025-01-23 - eg: assign variable & use of echo -e
+* 2025-01-21 - eg: use sed to do substitution [skip ci]
 * 2025-01-19 - eg: useful examples for awk [skip ci]
 * 2025-01-18 - eg: double brackets [] [skip ci]
 * 2025-01-17 - eg: test and brackets, common compares used in if condition [skip ci]